Separate player and recorder logics
