+++ src/components/sourceCast/SourcecastPlaybackControlbar.tsx
@@ -4,15 +4,12 @@ import { IconNames } from '@blueprintjs/icons';
 import * as React from 'react';
 
 import { controlButton } from '../commons';
-import { Recorder } from './util';
 
 class SourcecastPlaybackControlbar extends React.PureComponent<
   ISourcecastPlaybackControlbarProps,
   ISourcecastPlaybackControlbarState
 > {
   private audio: React.RefObject<HTMLAudioElement>;
-  private recorder: any;
-  private audioContext: AudioContext;
 
   constructor(props: ISourcecastPlaybackControlbarProps) {
     super(props);
@@ -21,37 +18,14 @@ class SourcecastPlaybackControlbar extends React.PureComponent<
       currentPlayerTime: 0,
       currentPlayerProgress: 0,
       duration: 0,
-      isPlayerMode: true,
-      fileDataBlob: null
+      isPlayerMode: true
     };
   }
 
-  public async componentDidMount() {
-    const constraints = { audio: true, video: false };
-    let getUserMediaNow: MediaStream;
-
-    try {
-      getUserMediaNow = await navigator.mediaDevices.getUserMedia(constraints);
-      this.startUserMedia(getUserMediaNow);
-      console.log(getUserMediaNow);
-    } catch (error) {
-      console.log('Microphone not found: ' + error);
-    }
-  }
-
   public render() {
-    const PlayerButton = controlButton('Player Mode', IconNames.MUSIC, this.handleSetPlayerMode);
-    const RecorderButton = controlButton(
-      'Recorder Mode',
-      IconNames.MOBILE_VIDEO,
-      this.handleSetRecorderMode
-    );
     const PlayerPlayButton = controlButton('Play', IconNames.PLAY, this.handlePlayerPlaying);
     const PlayerPauseButton = controlButton('Pause', IconNames.PAUSE, this.handlePlayerPausing);
     const PlayerStopButton = controlButton('Stop', IconNames.STOP, this.handlePlayerStopping);
-    const RecorderPlayButton = controlButton('Record', IconNames.PLAY, this.handleRecorderStarting);
-    const RecorderPauseButton = controlButton('Pause', IconNames.PAUSE, this.handleRecorderPausing);
-    const RecorderStopButton = controlButton('Stop', IconNames.STOP, this.handleRecorderStopping);
     return (
       <div>
         <audio
@@ -62,60 +36,30 @@ class SourcecastPlaybackControlbar extends React.PureComponent<
           preload="metadata"
           // controls={true}
         />
-        <div className="PlayerControl">
-          {PlayerButton}
-          {RecorderButton}
-        </div>
         <br />
-        {this.state.isPlayerMode ? (
-          <div>
-            <div className="Slider">
-              <Slider
-                min={0}
-                max={1}
-                stepSize={0.001}
-                onChange={this.handlePlayerProgressBarChange}
-                value={this.state.currentPlayerProgress}
-                labelRenderer={this.renderLabel}
-              />
-            </div>
-            <div className="PlayerControl">
-              {PlayerPlayButton}
-              {PlayerPauseButton}
-              {PlayerStopButton}
-            </div>
+        <div>
+          <div className="Slider">
+            <Slider
+              min={0}
+              max={1}
+              stepSize={0.001}
+              onChange={this.handlePlayerProgressBarChange}
+              value={this.state.currentPlayerProgress}
+              labelRenderer={this.renderLabel}
+            />
           </div>
-        ) : (
-          <div>
-            <div className="Slider">
-              <Slider min={0} max={1} stepSize={0.001} value={1} labelRenderer={this.renderLabel} />
-            </div>
-            <div className="PlayerControl">
-              {RecorderPlayButton}
-              {RecorderPauseButton}
-              {RecorderStopButton}
-            </div>
+          <div className="PlayerControl">
+            {PlayerPlayButton}
+            {PlayerPauseButton}
+            {PlayerStopButton}
           </div>
-        )}
+        </div>
       </div>
     );
   }
 
   private handleAudioLoaded = () => {
     this.props.handleSetSourcecastPlaybackDuration(this.audio.current!.duration);
-  };
-
-  private handleSetPlayerMode = () => {
-    this.setState({
-      isPlayerMode: true
-    });
-  };
-
-  private handleSetRecorderMode = () => {
-    this.handlePlayerStopping();
-    this.setState({
-      isPlayerMode: false
-    });
   };
 
   private handlePlayerPlaying = () => {
@@ -166,43 +110,6 @@ class SourcecastPlaybackControlbar extends React.PureComponent<
     }
   };
 
-  private startUserMedia = (stream: MediaStream) => {
-    this.audioContext = new AudioContext();
-    const input = this.audioContext.createMediaStreamSource(stream);
-    this.recorder = new Recorder(input);
-  };
-
-  private handleRecorderStarting = () => {
-    this.recorder.record();
-    console.log('Start recording');
-  };
-
-  private handleRecorderPausing = () => {
-    this.recorder.stop();
-    console.log('Pause recording');
-  };
-
-  private handleRecorderStopping = () => {
-    this.recorder.stop();
-    console.log('Stop recording');
-    this.createDownloadLink();
-    this.recorder.clear();
-  };
-
-  private createDownloadLink = () => {
-    this.recorder.exportWAV((blob: any) => {
-      console.log(blob);
-      this.setState({
-        fileDataBlob: blob
-      });
-      if (!blob) {
-        console.log('No recording');
-      } else {
-        console.log('Download link created: ' + URL.createObjectURL(blob));
-      }
-    });
-  };
-
   private renderLabel = (value: number) => {
     const totalTime = this.props.duration * value;
     const min = Math.floor(totalTime / 60);
@@ -227,7 +134,6 @@ export interface ISourcecastPlaybackControlbarState {
   currentPlayerProgress: number;
   duration: number;
   isPlayerMode: boolean;
-  fileDataBlob: any;
 }
 
 export default SourcecastPlaybackControlbar;
